+++ docs/fr_FR/changelog.md
@@ -1,8 +1,18 @@
-Beta
-####
+# Beta
 
 0.0.1 - 24 mars 2020
 ---
 
 * Gestion de détection d'inactivité
 * Création documentation
+
+0.0.2 - 29 mars
+---
+
+* Ajout de la possibilité d'un AR de l'alerte par les aidants
+* Ajout séquencement des actions d'alertes et lien entre elles pour ne déclencher les annulations que si l'action liée a été réalisée
+* Ajout des tags de configuration de la personne
+* Ajout gestion absence via plugin agenda ou capteurs+timer ou via appel externe (API ou autre plugin jeedom)
+* Ajout 1 timer de détection d'activité par capteur
+* Tests et debugs
+* Mise à jour documentation